+++ packages/insomnia/src/ui/components/request-url-bar.tsx
@@ -6,7 +6,7 @@ import { useParams, useSearchParams } from 'react-router';
 import * as reactUse from 'react-use';
 
 import { SECURITY_SETTINGS_PATH_LABEL } from '~/common/misc';
-import { recordProjectRecentRequest } from '~/common/project';
+import { recordProjectRecentRequest } from '~/ui/utils/recent-project-requests';
 import { useRootLoaderData } from '~/root';
 import {
   type ConnectActionParams,
