+++ packages/insomnia/src/ui/components/websockets/action-bar.tsx
@@ -3,7 +3,7 @@ import { services } from 'insomnia-data';
 import React, { forwardRef, useCallback, useEffect, useImperativeHandle, useLayoutEffect, useRef } from 'react';
 import { useParams } from 'react-router';
 
-import { recordProjectRecentRequest } from '~/common/project';
+import { recordProjectRecentRequest } from '~/ui/utils/recent-project-requests';
 import {
   type ConnectActionParams,
   useRequestConnectActionFetcher,
